Notification flicker fix  + duration field in media items
